split out vectors and curves into its own file
i also exported them instead of declaring them globally, so i updated some files
to import it
