+++ src/components/tabs/celestial-navigation/CelestialNavigationTab.vue
@@ -6,6 +6,7 @@ import NodeBackground from "./NodeBackground";
 import NodeOverlay from "./NodeOverlay";
 import NodeRing from "./NodeRing";
 import ProgressConnector from "./ProgressConnector";
+import { Vector } from "#utils/vectors.js";
 
 export default {
   name: "CelestialNavigationTab",
