+++ src/core/celestials/ra/alchemy.js
@@ -1,4 +1,5 @@
 import { ALCHEMY_RESOURCE, DC } from "#utils/constants.js";
+import { ExponentialMovingAverage } from "#utils/vectors.js";
 import { poissonDistribution } from "#utils/math.js";
 import { GameMechanicState } from "../../game-mechanics";
