+++ src/utils/math.js
@@ -1135,523 +1135,3 @@ window.testSSBQ = function testSSBQ(A, B, C, D, E, F) {
     console.log(`${s.x} ${s.y} ${e1} ${e2}`);
   }
 };
-
-window.AffineTransform = class AffineTransform {
-  constructor(a00 = 1, a01 = 0, a10 = 0, a11 = 1, o0 = 0, o1 = 0) {
-    this.a00 = a00;
-    this.a01 = a01;
-    this.a10 = a10;
-    this.a11 = a11;
-    this.o0 = o0;
-    this.o1 = o1;
-  }
-
-  times(ot) {
-    if (ot instanceof AffineTransform) {
-      return new AffineTransform(
-        this.a00 * ot.a00 + this.a01 * ot.a10, this.a00 * ot.a01 + this.a01 * ot.a11,
-        this.a10 * ot.a00 + this.a11 * ot.a10, this.a10 * ot.a01 + this.a11 * ot.a11,
-        this.a00 * ot.o0 + this.a01 * ot.o1 + this.o0,
-        this.a10 * ot.o0 + this.a11 * ot.o1 + this.o1,
-      );
-    }
-    if (ot instanceof Vector) return ot.transformedBy(this);
-    throw new Error("unsupported operation");
-  }
-
-  translated(offX, offY) {
-    if (offX instanceof Vector) {
-      return new AffineTransform(this.a00, this.a01, this.a10, this.a11, this.o0 + offX.x, this.o1 + offX.y);
-    }
-    return new AffineTransform(this.a00, this.a01, this.a10, this.a11, this.o0 + offX, this.o1 + offY);
-  }
-
-  rotated(angle) {
-    return AffineTransform.rotation(angle).times(this);
-  }
-
-  scaled(scale) {
-    return AffineTransform.scale(scale).times(this);
-  }
-
-  get withoutTranslation() {
-    return new AffineTransform(this.a00, this.a01, this.a10, this.a11);
-  }
-
-  static translation(offX, offY) {
-    if (offX instanceof Vector) {
-      return new AffineTransform(1, 0, 0, 1, offX.x, offX.y);
-    }
-    return new AffineTransform(1, 0, 0, 1, offX, offY);
-  }
-
-  static rotation(angle) {
-    const c = Math.cos(angle), s = Math.sin(angle);
-    return new AffineTransform(c, -s, s, c);
-  }
-
-  static scale(sc) {
-    return new AffineTransform(sc, 0, 0, sc);
-  }
-
-  static identity() {
-    return new AffineTransform();
-  }
-};
-
-window.Vector = class Vector {
-  constructor(x, y) {
-    this.x = x;
-    this.y = y;
-  }
-
-  get length() {
-    return Math.sqrt(this.x * this.x + this.y * this.y);
-  }
-
-  plus(v) {
-    return new Vector(this.x + v.x, this.y + v.y);
-  }
-
-  dot(v) {
-    return this.x * v.x + this.y * v.y;
-  }
-
-  cross(v) {
-    // Produces scalar, z term of 3D vectors
-    return this.x * v.y - this.y * v.x;
-  }
-
-  minus(v) {
-    return new Vector(this.x - v.x, this.y - v.y);
-  }
-
-  times(s) {
-    return new Vector(this.x * s, this.y * s);
-  }
-
-  asTranslate() {
-    return `translate(${this.x}, ${this.y})`;
-  }
-
-  asRotate() {
-    return `rotate(${180 / Math.PI * Math.atan2(this.y, this.x)})`;
-  }
-
-  toString() {
-    return `${this.x}, ${this.y}`;
-  }
-
-  round(factor) {
-    return new Vector(Math.round(this.x * factor) / factor, Math.round(this.y * factor) / factor);
-  }
-
-  get copy() {
-    return new Vector(this.x, this.y);
-  }
-
-  matrixTransform(a00, a01, a10, a11) {
-    return new Vector(a00 * this.x + a01 * this.y, a10 * this.x + a11 * this.y);
-  }
-
-  transformedBy(tform) {
-    return new Vector(tform.a00 * this.x + tform.a01 * this.y + tform.o0,
-      tform.a10 * this.x + tform.a11 * this.y + tform.o1);
-  }
-
-  get negative() {
-    return new Vector(-this.x, -this.y);
-  }
-
-  get normalized() {
-    return this.times(1 / this.length);
-  }
-
-  get right90() {
-    return new Vector(this.y, -this.x);
-  }
-
-  get left90() {
-    return new Vector(-this.y, this.x);
-  }
-
-  get angle() {
-    return Math.atan2(this.y, this.x);
-  }
-
-  static horiz(x) {
-    return new Vector(x, 0);
-  }
-
-  static unitFromRadians(rad) {
-    return new Vector(Math.cos(rad), Math.sin(rad));
-  }
-
-  static unitFromDegrees(deg) {
-    return Vector.unitFromRadians(deg * Math.PI / 180);
-  }
-};
-
-window.Curve = class Curve {
-  /**
-   * @abstract
-   * @param {number} t
-   * @returns {Vector}
-  */
-  position() {
-    throw new NotImplementedError();
-  }
-
-  /**
-   * @abstract
-   * @param {number} t
-   * @returns {Vector}
-  */
-  derivative() {
-    throw new NotImplementedError();
-  }
-
-  /**
-   * @abstract
-   * @param {number} t
-   * @returns {Vector}
-  */
-  secondDerivative() {
-    throw new NotImplementedError();
-  }
-
-  /**
-   * @param {number} t
-   * @returns {number}
-   */
-  curvature(t) {
-    const d = this.derivative(t);
-    const dd = this.secondDerivative(t);
-    const dMag = d.length;
-    return d.cross(dd) / (dMag * dMag * dMag);
-  }
-
-  shapeAt(t) {
-    const d = this.derivative(t);
-    return {
-      t,
-      position: this.position(t),
-      derivative: d,
-      direction: d.normalized,
-      curvature: this.curvature(t),
-    };
-  }
-
-  minimumDistanceTo(pDes, tMin, tMax) {
-    let tGuess = 0.5 * (tMin + tMax);
-    const tTol = Math.max(Math.abs(tMax), Math.abs(tMin)) * Number.EPSILON * 16;
-    for (let iter = 0; ; ++iter) {
-      const p = this.position(tGuess);
-      const d = this.derivative(tGuess);
-      const dd = this.secondDerivative(tGuess);
-      const offset = p.minus(pDes);
-      const dist = offset.length;
-      const distDeriv = offset.dot(d) * 2;
-
-      if (distDeriv > 0) tMax = tGuess;
-      else tMin = tGuess;
-
-      const distSecondDeriv = (offset.dot(dd) + d.dot(d)) * 2;
-      const tStep = distSecondDeriv < 0 ? -dist / distDeriv : -distDeriv / distSecondDeriv;
-      if (Math.abs(tStep) < tTol || iter >= 16) return dist;
-      tGuess = Math.clamp(tGuess + tStep, tMin, tMax);
-    }
-  }
-};
-
-window.LinearPath = class LinearPath extends Curve {
-  constructor(p0, p1) {
-    super();
-    this.p0 = p0.copy;
-    this.p1 = p1.copy;
-  }
-
-  position(t) {
-    return this.p0.times(1 - t).plus(this.p1.times(t));
-  }
-
-  derivative() {
-    return this.p1.minus(this.p0);
-  }
-
-  secondDerivative() {
-    return new Vector(0, 0);
-  }
-
-  curvature(t) {
-    return 0;
-  }
-
-  trimStart(len) {
-    const dir = this.p1.minus(this.p0).normalized;
-    return new LinearPath(this.p0.plus(dir.times(len)), this.p1);
-  }
-
-  trimEnd(len) {
-    const dir = this.p1.minus(this.p0).normalized;
-    return new LinearPath(this.p0, this.p1.minus(dir.times(len)));
-  }
-
-  transformed(tform) {
-    return new LinearPath(this.p0.transformedBy(tform), this.p1.transformedBy(tform));
-  }
-
-  get relativeSVG() {
-    const d1 = this.p1.minus(this.p0);
-    return `l ${d1.x} ${d1.y}\n`;
-  }
-
-  createOffsetLine(offset, t0 = 0, t1 = 1) {
-    const off = this.p1.minus(this.p0).normalized.right90.times(offset);
-    return new LinearPath(this.position(t0).plus(off), this.position(t1).plus(off));
-  }
-
-  static connectCircles(p0, r0, p1, r1) {
-    const dir = p1.minus(p0).normalized;
-    return new LinearPath(p0.plus(dir.times(r0)), p1.minus(dir.times(r1)));
-  }
-};
-
-class CubicBezier extends Curve {
-  constructor(p0, p1, p2, p3) {
-    super();
-    this.p0 = p0.copy;
-    this.p1 = p1.copy;
-    this.p2 = p2.copy;
-    this.p3 = p3.copy;
-  }
-
-  position(t) {
-    const nt2 = (1 - t) * (1 - t);
-    const t2 = t * t;
-    return this.p0.times((1 - t) * nt2)
-      .plus(this.p1.times(3 * t * nt2))
-      .plus(this.p2.times(3 * t2 * (1 - t)))
-      .plus(this.p3.times(t2 * t));
-  }
-
-  derivative(t) {
-    return this.p1.minus(this.p0).times(3 * (1 - t) * (1 - t))
-      .plus(this.p2.minus(this.p1).times(6 * t * (1 - t)))
-      .plus(this.p3.minus(this.p2).times(3 * t * t));
-  }
-
-  secondDerivative(t) {
-    return this.p2.minus(this.p1.times(2)).plus(this.p0).times(6 * (1 - t))
-      .plus(this.p3.minus(this.p2.times(2)).plus(this.p1).times(6 * t));
-  }
-
-  transformed(tform) {
-    return new CubicBezier(this.p0.transformedBy(tform), this.p1.transformedBy(tform),
-      this.p2.transformedBy(tform), this.p3.transformedBy(tform));
-  }
-
-  get relativeSVG() {
-    const d1 = this.p1.minus(this.p0);
-    const d2 = this.p2.minus(this.p0);
-    const d3 = this.p3.minus(this.p0);
-    return `c ${d1.x} ${d1.y} ${d2.x} ${d2.y} ${d3.x} ${d3.y}\n`;
-  }
-
-  get reverse() {
-    return new CubicBezier(this.p3, this.p2, this.p1, this.p0);
-  }
-
-  static fitCurveSection(shape0, shape1) {
-    const dP = shape1.position.minus(shape0.position);
-    const reversed = shape0.t > shape1.t;
-    const pathRotation = shape0.direction.cross(shape1.direction);
-    let magSol = solveSimpleBiquadratic(
-      1.5 * shape0.curvature, pathRotation, -shape0.direction.cross(dP),
-      1.5 * shape1.curvature, pathRotation, shape1.direction.cross(dP));
-    magSol = reversed ? magSol.filter(o => o.x <= 0 && o.y <= 0) : magSol.filter(o => o.x >= 0 && o.y >= 0);
-    if (magSol.length === 0) return null;
-    return new CubicBezier(
-      shape0.position, shape0.position.plus(shape0.direction.times(magSol[0].x)),
-      shape1.position.minus(shape1.direction.times(magSol[0].y)), shape1.position);
-  }
-}
-
-// This is an "inset/outset" kind of transform
-window.OffsetCurve = class OffsetCurve extends Curve {
-  constructor(baseCurve, offset) {
-    super();
-    this.base = baseCurve;
-    this.offset = offset;
-  }
-
-  position(t) {
-    const p = this.base.position(t);
-    const d = this.base.derivative(t);
-    return p.plus(d.normalized.right90.times(this.offset));
-  }
-
-  derivative(t) {
-    return this.base.derivative(t);
-  }
-
-  // 2nd derivative not implemented as only curvature is used atm
-  curvature(t) {
-    const c = this.base.curvature(t);
-    return 1 / (1 / c + this.offset);
-  }
-
-  shapeAt(t) {
-    const shape = this.base.shapeAt(t);
-    return {
-      t: shape.t,
-      position: shape.position.plus(shape.direction.right90.times(this.offset)),
-      derivative: shape.derivative,
-      direction: shape.direction,
-      curvature: shape.curvature / (1 + this.offset * shape.curvature),
-    };
-  }
-};
-
-window.LogarithmicSpiral = class LogarithmicSpiral extends Curve {
-  constructor(center, scale, rate) {
-    super();
-    this.center = center;
-    this.scale = scale;
-    this.rate = rate;
-  }
-
-  position(t) {
-    return Vector.unitFromRadians(t)
-      .times(this.scale * Math.exp(this.rate * t))
-      .plus(this.center);
-  }
-
-  derivative(t) {
-    const unit = Vector.unitFromRadians(t);
-    const radius = this.scale * Math.exp(this.rate * t);
-    return unit.times(radius * this.rate).plus(unit.left90.times(radius));
-  }
-
-  secondDerivative(t) {
-    const unit = Vector.unitFromRadians(t);
-    const radius = this.scale * Math.exp(this.rate * t);
-    return unit.times(radius * (this.rate * this.rate - 1))
-      .plus(unit.left90.times(2 * radius * this.rate));
-  }
-
-  shapeAt(t) {
-    const unit = Vector.unitFromRadians(t);
-    const radius = this.scale * Math.exp(this.rate * t);
-    const ur = unit.times(radius);
-    const d = ur.times(this.rate).plus(ur.left90);
-    return {
-      t,
-      position: ur.plus(this.center),
-      derivative: d,
-      direction: d.normalized,
-      curvature: 1 / (Math.abs(radius) * Math.sqrt(1 + this.rate * this.rate)),
-    };
-  }
-
-  angleFromRadius(r) {
-    return Math.log(r / this.scale) / this.rate;
-  }
-
-  static fromPolarEndpoints(center, theta0, r0, theta1, r1) {
-    const rate = Math.log(r1 / r0) / (theta1 - theta0);
-    return new LogarithmicSpiral(center, r0 / Math.exp(rate * theta0), rate);
-  }
-};
-
-window.PiecewisePath = class PiecewisePath {
-  constructor(data) {
-    this.path = data ? data : [];
-  }
-
-  push(element) {
-    this.path.push(element);
-  }
-
-  transformedBy(tform) {
-    return new PiecewisePath(this.path.map(x => x.transformed(tform)));
-  }
-
-  toSVG(initialPrefix) {
-    const p0 = this.path[0].position(0);
-    const lines = [`${initialPrefix} ${p0.x} ${p0.y}\n`];
-    for (const part of this.path) lines.push(part.relativeSVG);
-    return lines.join("");
-  }
-
-  static cubicBezierFitToCurveSection(curve, t0, t1, tol = 1, minPieces = 1) {
-    const output = new PiecewisePath();
-    const shape0 = curve.shapeAt(t0);
-    const shape1 = curve.shapeAt(t1);
-    function subdivide(shapeStart, shapeEnd, maxDepth = 8) {
-      const shapeMid = curve.shapeAt(0.5 * (shapeStart.t + shapeEnd.t));
-      return single(shapeStart, shapeMid, maxDepth - 1)
-        && single(shapeMid, shapeEnd, maxDepth - 1);
-    }
-    function single(shapeStart, shapeEnd, maxDepth = 8) {
-      const singleFit = CubicBezier.fitCurveSection(shapeStart, shapeEnd);
-      if (singleFit === null) {
-        if (maxDepth <= 0) throw new Error("coulnd't decompose curve");
-        return subdivide(shapeStart, shapeEnd, maxDepth);
-      }
-      const tMid = 0.5 * (shapeStart.t + shapeEnd.t);
-      const err = singleFit.minimumDistanceTo(curve.position(tMid), 0, 1);
-      if (err > tol) {
-        return subdivide(shapeStart, shapeEnd, maxDepth);
-      }
-      output.push(singleFit);
-      return true;
-    }
-    if (minPieces > 1) subdivide(shape0, shape1);
-    else single(shape0, shape1);
-    return output;
-  }
-};
-
-// https://stackoverflow.com/a/9201081
-window.ExponentialMovingAverage = class ExponentialMovingAverage {
-  constructor(alpha = 0.02, maxOutliers = 5, highOutlierThreshold = 3, lowOutlierThreshold = 0.4) {
-    this.alpha = alpha;
-    this.maxOutliers = maxOutliers;
-    this.highOutlierThreshold = highOutlierThreshold;
-    this.lowOutlierThreshold = lowOutlierThreshold;
-    this.outliers = 0;
-    this._average = new Decimal(0);
-  }
-
-  get average() {
-    if (this._average.eq(0)) {
-      return 0;
-    }
-    return this._average;
-  }
-
-  addValue(value) {
-    if (this._average.eq(0)) {
-      this._average = value;
-    } else {
-      this._average = this._average.add((value.sub(this._average)).mul(this.average));
-
-      const absValue = Decimal.abs(value);
-      const absAverage = Decimal.abs(this._average);
-      const highOutlier = absValue.gt(absAverage.mul(this.highOutlierThreshold));
-      const lowOutlier = absValue.lt(absAverage.mul(this.lowOutlierThreshold));
-      const outlier = highOutlier || lowOutlier;
-
-      if (outlier) {
-        this.outliers++;
-        if (this.outliers >= this.maxOutliers) {
-          this._average = value;
-          this.outliers = 0;
-        }
-      } else {
-        this.outliers = 0;
-      }
-    }
-  }
-};
